requirements.txt
click==6.7
Flask==1.1.2
Flask-Cors==3.0.8
Flask-HTTPAuth==4.0.0
Flask-SQLAlchemy==2.4.1
itsdangerous==0.24
Jinja2==2.11.2
MarkupSafe==1.0
PyJWT==1.7.1
SQLAlchemy==1.3.20
Werkzeug==1.0.1
requests==2.20.0
python-logstash
PyMySQL==0.10.1